planner: split index test
