+++ planner/core/casetest/index/BUILD.bazel
@@ -9,13 +9,14 @@ go_test(
     ],
     data = glob(["testdata/**"]),
     flaky = True,
-    shard_count = 8,
+    shard_count = 9,
     deps = [
         "//sessionctx/variable",
         "//testkit",
         "//testkit/testdata",
         "//testkit/testmain",
         "//testkit/testsetup",
+        "//util",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
