+++ planner/core/casetest/integration_test.go
@@ -27,7 +27,6 @@ import (
 	"github.com/pingcap/tidb/statistics/handle"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/pingcap/tidb/testkit/testdata"
-	"github.com/pingcap/tidb/util"
 	"github.com/stretchr/testify/require"
 )
 
@@ -1244,62 +1243,6 @@ func TestDowncastPointGetOrRangeScan(t *testing.T) {
 		tk.MustQuery("explain format='brief' " + tt).Check(testkit.Rows(output[i].Plan...))
 		tk.MustQuery(tt).Sort().Check(testkit.Rows(output[i].Result...))
 	}
-}
-
-func TestNullConditionForPrefixIndex(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec(`CREATE TABLE t1 (
-  id char(1) DEFAULT NULL,
-  c1 varchar(255) DEFAULT NULL,
-  c2 text DEFAULT NULL,
-  KEY idx1 (c1),
-  KEY idx2 (c1,c2(5))
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin`)
-	tk.MustExec("set tidb_cost_model_version=2")
-	tk.MustExec("create table t2(a int, b varchar(10), index idx(b(5)))")
-	tk.MustExec("create table t3(a int, b varchar(10), c int, primary key (a, b(5)) clustered)")
-	tk.MustExec("set tidb_opt_prefix_index_single_scan = 1")
-	tk.MustExec("insert into t1 values ('a', '0xfff', '111111'), ('b', '0xfff', '22    '), ('c', '0xfff', ''), ('d', '0xfff', null)")
-	tk.MustExec("insert into t2 values (1, 'aaaaaa'), (2, 'bb    '), (3, ''), (4, null)")
-	tk.MustExec("insert into t3 values (1, 'aaaaaa', 2), (1, 'bb    ', 3), (1, '', 4)")
-
-	var input []string
-	var output []struct {
-		SQL    string
-		Plan   []string
-		Result []string
-	}
-	integrationSuiteData := GetIntegrationSuiteData()
-	integrationSuiteData.LoadTestCases(t, &input, &output)
-	for i, tt := range input {
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Plan = testdata.ConvertRowsToStrings(tk.MustQuery("explain format='brief' " + tt).Rows())
-			output[i].Result = testdata.ConvertRowsToStrings(tk.MustQuery(tt).Sort().Rows())
-		})
-		tk.MustQuery("explain format='brief' " + tt).Check(testkit.Rows(output[i].Plan...))
-		tk.MustQuery(tt).Sort().Check(testkit.Rows(output[i].Result...))
-	}
-
-	// test plan cache
-	tk.MustExec(`set tidb_enable_prepared_plan_cache=1`)
-	tk.MustExec("set @@tidb_enable_collect_execution_info=0")
-	tk.MustExec("prepare stmt from 'select count(1) from t1 where c1 = ? and c2 is not null'")
-	tk.MustExec("set @a = '0xfff'")
-	tk.MustQuery("execute stmt using @a").Check(testkit.Rows("3"))
-	tk.MustQuery("execute stmt using @a").Check(testkit.Rows("3"))
-	tk.MustQuery(`select @@last_plan_from_cache`).Check(testkit.Rows("1"))
-	tk.MustQuery("execute stmt using @a").Check(testkit.Rows("3"))
-	tkProcess := tk.Session().ShowProcess()
-	ps := []*util.ProcessInfo{tkProcess}
-	tk.Session().SetSessionManager(&testkit.MockSessionManager{PS: ps})
-	tk.MustQuery(fmt.Sprintf("explain for connection %d", tkProcess.ID)).Check(testkit.Rows(
-		"StreamAgg_17 1.00 root  funcs:count(Column#7)->Column#5",
-		"└─IndexReader_18 1.00 root  index:StreamAgg_9",
-		"  └─StreamAgg_9 1.00 cop[tikv]  funcs:count(1)->Column#7",
-		"    └─IndexRangeScan_16 99.90 cop[tikv] table:t1, index:idx2(c1, c2) range:[\"0xfff\" -inf,\"0xfff\" +inf], keep order:false, stats:pseudo"))
 }
 
 // https://github.com/pingcap/tidb/issues/24095
